Parameter getter/setter

+++ src/SimpleNumConverter.ipp
@@ -79,7 +79,7 @@ Poco::Int32 SimpleNumConverter::getInt32(const float& data)
 	else if (data < std::numeric_limits<Poco::Int32>::min())
 		return std::numeric_limits<Poco::Int32>::min();
 	else
-		return static_cast<Poco::Int32>(round(data));
+		return static_cast<Poco::Int32>(round(static_cast<double>(data)));
 }
 
 template <> inline
@@ -90,7 +90,7 @@ Poco::Int32 SimpleNumConverter::getInt32(const double& data)
 	else if (data < std::numeric_limits<Poco::Int32>::min())
 		return std::numeric_limits<Poco::Int32>::min();
 	else
-		return static_cast<Poco::Int32>(round(data));
+		return static_cast<Poco::Int32>(round(static_cast<double>(data)));
 }
 
 template<typename T>
@@ -140,7 +140,7 @@ Poco::UInt32 SimpleNumConverter::getUInt32(const float& data)
 	else if (data > std::numeric_limits<Poco::UInt32>::max())
 		return std::numeric_limits<Poco::UInt32>::max();
 	else
-		return static_cast<Poco::UInt32>(round(data));
+		return static_cast<Poco::UInt32>(round(static_cast<double>(data)));
 }
 
 template <> inline
@@ -151,7 +151,7 @@ Poco::UInt32 SimpleNumConverter::getUInt32(const double& data)
 	else if (data > std::numeric_limits<Poco::UInt32>::max())
 		return std::numeric_limits<Poco::UInt32>::max();
 	else
-		return static_cast<Poco::UInt32>(round(data));
+		return static_cast<Poco::UInt32>(round(static_cast<double>(data)));
 }
 
 template<typename T>
@@ -170,7 +170,7 @@ Poco::Int64 SimpleNumConverter::getInt64(const Poco::UInt64& data)
 	if (data > std::numeric_limits<Poco::Int64>::max())
 		return std::numeric_limits<Poco::Int64>::max();
 	else
-		return static_cast<Poco::Int64>(round(data));
+		return static_cast<Poco::Int64>(round(static_cast<double>(data)));
 }
 
 template <> inline
@@ -181,7 +181,7 @@ Poco::Int64 SimpleNumConverter::getInt64(const float& data)
 	else if (data < std::numeric_limits<Poco::Int64>::min())
 		return std::numeric_limits<Poco::Int64>::min();
 	else
-		return static_cast<Poco::Int64>(round(data));
+		return static_cast<Poco::Int64>(round(static_cast<double>(data)));
 }
 
 template <> inline
@@ -192,7 +192,7 @@ Poco::Int64 SimpleNumConverter::getInt64(const double& data)
 	else if (data < std::numeric_limits<Poco::Int64>::min())
 		return std::numeric_limits<Poco::Int64>::min();
 	else
-		return static_cast<Poco::Int64>(round(data));
+		return static_cast<Poco::Int64>(round(static_cast<double>(data)));
 }
 
 template<typename T>
@@ -231,7 +231,7 @@ Poco::UInt64 SimpleNumConverter::getUInt64(const float& data)
 	else if (data > std::numeric_limits<Poco::UInt64>::max())
 		return std::numeric_limits<Poco::UInt64>::max();
 	else
-		return static_cast<Poco::UInt64>(round(data));
+		return static_cast<Poco::UInt64>(round(static_cast<double>(data)));
 }
 
 template <> inline
@@ -242,7 +242,7 @@ Poco::UInt64 SimpleNumConverter::getUInt64(const double& data)
 	else if (data > std::numeric_limits<Poco::UInt64>::max())
 		return std::numeric_limits<Poco::UInt64>::max();
 	else
-		return static_cast<Poco::UInt64>(round(data));
+		return static_cast<Poco::UInt64>(round(static_cast<double>(data)));
 }
 
 template<typename T>
